requirements.txt
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=0.24.0
tensorflow>=2.4.0
pymongo>=3.11.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
ta-lib>=0.4.0
pytest>=6.2.0
flake8>=3.9.0
ccxt~=4.3.72
setuptools~=72.1.0
scipy~=1.14.0